Remove Int64 (#334)

+++ src/accelerated_gradient_descent.jl
@@ -26,7 +26,7 @@ type AcceleratedGradientDescentState{T}
     x_previous::Array{T}
     g::Array{T}
     f_x_previous::T
-    iteration::Int64
+    iteration::Int
     y::Array{T}
     y_previous::Array{T}
     s::Array{T}
